requirements.txt
# 深度学习框架
torch>=2.1.0
# 数值与表格处理
numpy>=1.24.0
pandas>=2.0.0
scipy>=1.11.0
# 配置与日志
pyyaml>=6.0
tqdm>=4.66.0
# EEG / fMRI 数据读取与 ROI 提取
mne>=1.6.0
nibabel>=5.2.0
nilearn>=0.10.0
